Lock file in a deferred manner before sending to processor
DirectoryWatcher now has to queues - immediate and deferred. Some events can be
immediate, e.g file remove. Events, which result in an existing file (RenameNew,
Add, Modify) have to be locked, because they may still be used. Those events
will be pushed to the deferred queue.

Deferred queue is processed by DeferredFileLocker which tries to lock all the
files it has. When it succeeds with either of them, it pushes it to the actual
queue that is consumed by the Processor.

We still unlock file before processing so there's a race condition here, though
very improbable, so we're leaving this for now.