Fix acceptance tests
We had a race condition when waiting for Charon to start and stop.
Instead of some artificial timeouts, we should wait for certain stdout
messages to appear in stdout. To be able to do this, I added a verbose
mode to Charon, so it can notify that filesystem operation has been
completed.